requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Admin==1.6.1
Werkzeug==3.0.1
pandas==2.1.4
twilio==8.10.0
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-Mail==0.9.1
WTForms==3.0.1
requests==2.32.5
psutil==5.9.6
